+++ changelog.md
@@ -5,11 +5,12 @@
 ### Major
 
 - Updated Node.js support to `^12.20 || >= 14.13`.
+- Updated the [`graphql-react`](https://npm.im/graphql-react) peer dependency to `^14.0.0`.
+- Updated dependencies, some of which require newer Node.js versions than were previously supported.
 - Replaced the the `package.json` `exports` field public [subpath folder mapping](https://nodejs.org/api/packages.html#packages_subpath_folder_mappings) (deprecated by Node.js) with a [subpath pattern](https://nodejs.org/api/packages.html#packages_subpath_patterns). Deep `require` paths within `next-graphql-react/public/` must now include the `.js` file extension.
 
 ### Patch
 
-- Updated dev dependencies.
 - Updated GitHub Actions CI config to run tests with Node.js v12, v14, v16.
 - Updated the [example Next.js app](https://graphql-react.vercel.app) URL in the readme.